TestProject1/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2023.3
requests==2.28.2
six==1.16.0
urllib3==1.26.15
XlsxWriter==3.0.9
